+++ notebooker/__init__.py
@@ -1,3 +1,5 @@
+from pkgutil import extend_path
+
 from .version import __version__
 
-__import__("pkg_resources").declare_namespace(__name__)
+__path__ = extend_path(__path__, __name__)